orchestration/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
scrapy==2.11.0
apache-airflow-providers-microsoft-azure==4.0.0
psycopg2-binary